% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood.R
\name{calories_in_Kg}
\alias{calories_in_Kg}
\title{Calories conversion}
\usage{
calories_in_Kg(calories)
}
\arguments{
\item{calories}{A numeric calorie number in grams to be converted to Kilograms}
}
\value{
the converted calorie value in Kgs
}
\description{
Converts a given calorie value(in grams) of a fast food item to Kilograms(Kgs)
}
\examples{
calories_in_Kg(38)

}
