% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood.R
\name{cholesterol}
\alias{cholesterol}
\title{Highest Cholesterol}
\usage{
cholesterol(restaurant)
}
\arguments{
\item{restaurant}{A character argument which implies a restaurant}
}
\value{
a vector
}
\description{
Prints out a list of items with highest cholesterol content based on the restaurant selected
}
\examples{
cholesterol()
cholesterol(Mcdonalds)

}
