% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastfoodinfo}
\alias{fastfoodinfo}
\title{Nutritional content of Fast food items}
\format{
A data frame with  9287 observations on 18 variables. The variables are as follows :

\itemize{
\item	\strong{restaurant}	Name of fast food restaurant
\item	\strong{item}	Fast food item from that restaurant
\item	\strong{calories}	Calories in Kilo Joules
\item	\strong{cal_fat}	calories from fat in Kilo Joules
\item	\strong{total_fat}	Total fat in grams
\item	\strong{sat_fat}	saturated fat in grams
\item	\strong{trans_fat}	trans-unsaturated fats
\item	\strong{cholesterol}	Cholesterol in milli-grams
\item	\strong{Sodium} Sodium content in milli-grams
\item	\strong{total_carb}	total Number of carbohydrates in grams
\item	\strong{fiber} Fiber content in grams
\item	\strong{sugar}	Sugar content in grams
\item	\strong{protein}	Protein content in grams
\item	\strong{vit_a} Vitamin A quantity in grams
\item  \strong{vit_c} Vitamin C quantity in grams
\item	\strong{calcium} Calcium quantity in grams
\item	\strong{salad}	salad quanitity(misc)
}
}
\source{
\url{https://fastfoodnutrition.org/}
}
\usage{
data(fastfoodinfo)
}
\description{
This data is taken from the \emph{Fast Food Nutrition} website and comprises of 8 famous fast food chains along with their items and gives a nutritional breakdown of these contents
}
\keyword{datasets}
