% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfood.R
\name{sugar_content}
\alias{sugar_content}
\title{Sugar Quantity}
\usage{
sugar_content(sugar)
}
\arguments{
\item{sugar}{A numeric argument that denotes a sugar quantity of a fast food item}
}
\value{
a string that prints a message based on if the function condition is satisfied
}
\description{
Prints a message based on the sugar quantity entered
}
\examples{
sugar_content()
sugar_content(5)
sugar_content(14)

}
